Add getDisplay implementation for R.
Returns the default shadow display. This method in the real implementation throws if the context is not a "UI context", but in Robolectric the same ContextImpl is used for activities and the application. However it's more useful for the shadow implementation to return the default display as this method will typically only be invoked within the correct context (as it throws otherwise).

PiperOrigin-RevId: 374414615